Migrate solve.py into respective detect/ modules
