+++ pooltool/evolution/event_based/detect/ball_ball.py
@@ -3,13 +3,120 @@ from __future__ import annotations
 from itertools import combinations
 
 import numpy as np
+from numba import jit
+from numpy.typing import NDArray
 
 import pooltool.constants as const
 import pooltool.ptmath as ptmath
 from pooltool.events import Event, EventType, ball_ball_collision, null_event
 from pooltool.evolution.event_based.cache import CollisionCache
-from pooltool.physics.motion.solve import ball_ball_collision_time
+from pooltool.physics.utils import get_u_vec
+from pooltool.ptmath.roots import quartic
+from pooltool.ptmath.roots.core import get_real_positive_smallest_root
 from pooltool.system.datatypes import System
+
+
+@jit(nopython=True, cache=const.use_numba_cache)
+def ball_ball_collision_coeffs(
+    rvw1: NDArray[np.float64],
+    rvw2: NDArray[np.float64],
+    s1: int,
+    s2: int,
+    mu1: float,
+    mu2: float,
+    m1: float,
+    m2: float,
+    g1: float,
+    g2: float,
+    R: float,
+) -> tuple[float, float, float, float, float]:
+    """Get quartic coeffs required to determine the ball-ball collision time
+
+    (just-in-time compiled)
+    """
+
+    c1x, c1y = rvw1[0, 0], rvw1[0, 1]
+    c2x, c2y = rvw2[0, 0], rvw2[0, 1]
+
+    if s1 == const.spinning or s1 == const.pocketed or s1 == const.stationary:
+        a1x, a1y, b1x, b1y = 0, 0, 0, 0
+    else:
+        phi1 = ptmath.angle(rvw1[1])
+        v1 = ptmath.norm3d(rvw1[1])
+
+        u1 = get_u_vec(rvw1, R, phi1, s1)
+
+        K1 = -0.5 * mu1 * g1
+        cos_phi1 = np.cos(phi1)
+        sin_phi1 = np.sin(phi1)
+
+        a1x = K1 * (u1[0] * cos_phi1 - u1[1] * sin_phi1)
+        a1y = K1 * (u1[0] * sin_phi1 + u1[1] * cos_phi1)
+        b1x = v1 * cos_phi1
+        b1y = v1 * sin_phi1
+
+    if s2 == const.spinning or s2 == const.pocketed or s2 == const.stationary:
+        a2x, a2y, b2x, b2y = 0.0, 0.0, 0.0, 0.0
+    else:
+        phi2 = ptmath.angle(rvw2[1])
+        v2 = ptmath.norm3d(rvw2[1])
+
+        u2 = get_u_vec(rvw2, R, phi2, s2)
+
+        K2 = -0.5 * mu2 * g2
+        cos_phi2 = np.cos(phi2)
+        sin_phi2 = np.sin(phi2)
+
+        a2x = K2 * (u2[0] * cos_phi2 - u2[1] * sin_phi2)
+        a2y = K2 * (u2[0] * sin_phi2 + u2[1] * cos_phi2)
+        b2x = v2 * cos_phi2
+        b2y = v2 * sin_phi2
+
+    Ax, Ay = a2x - a1x, a2y - a1y
+    Bx, By = b2x - b1x, b2y - b1y
+    Cx, Cy = c2x - c1x, c2y - c1y
+
+    a = Ax * Ax + Ay * Ay
+    b = 2 * Ax * Bx + 2 * Ay * By
+    c = Bx * Bx + 2 * Ax * Cx + 2 * Ay * Cy + By * By
+    d = 2 * Bx * Cx + 2 * By * Cy
+    e = Cx * Cx + Cy * Cy - 4 * R * R
+
+    return a, b, c, d, e
+
+
+@jit(nopython=True, cache=const.use_numba_cache)
+def ball_ball_collision_time(
+    rvw1: NDArray[np.float64],
+    rvw2: NDArray[np.float64],
+    s1: int,
+    s2: int,
+    mu1: float,
+    mu2: float,
+    m1: float,
+    m2: float,
+    g1: float,
+    g2: float,
+    R: float,
+) -> float:
+    """Get the time until collision between 2 balls."""
+    return get_real_positive_smallest_root(
+        quartic.solve(
+            *ball_ball_collision_coeffs(
+                rvw1,
+                rvw2,
+                s1,
+                s2,
+                mu1,
+                mu2,
+                m1,
+                m2,
+                g1,
+                g2,
+                R,
+            )
+        )
+    )
 
 
 def get_next_ball_ball_2d_event(shot: System, collision_cache: CollisionCache) -> Event:
