+++ tests/evolution/event_based/test_simulate.py
@@ -13,11 +13,11 @@ from pooltool.evolution.event_based.detect import (
     EventDetector,
     get_next_ball_ball_2d_event,
 )
+from pooltool.evolution.event_based.detect.ball_ball import ball_ball_collision_time
 from pooltool.evolution.event_based.simulate import simulate
 from pooltool.objects import Ball, BilliardTableSpecs, Cue, Table
 from pooltool.objects.ball.params import BallParams
 from pooltool.objects.ball.sets import BallSet
-from pooltool.physics.motion.solve import ball_ball_collision_time
 from pooltool.ptmath.roots import quadratic
 from pooltool.system import System
 from tests.evolution.event_based.test_data import TEST_DIR
